moved test server inside src folder
